pyproject.toml
[tool.poetry]
name = "foto-para-desenho"
version = "0.1.0"
description = ""
authors = ["Ricardo Roson <rroson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
streamlit = "^1.38.0"
pillow = "^10.4.0"
numpy = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
